Improve undo buffer and StateFn protocol change
At the cost of one undo level, we can undo the first byte and report an
invalid position.

All StateFn, including init, expect the first rune to be read with
State.Current() and upon return, State.Current() must be the first
non-matched rune in the previous iteration. The drawback is that bugs in
StateFn often end-up in endless loops. Add a check in init handler?

Need to benchmark different ring buffer algorithms